Schedule and Speakers Updated
I added in the schedule details to the schedule page. The speaker page is also started being formatted correctly to flex across the page. Talked to Tim about creating my .mp4 to a gif when it changes to mobile size. Additional work and formatting was added to the registration page as well like including buttons.