Code of Conduct, Registration, and Sponsors
The code of conduct page, about page, sponsor page and registration pages were created and styled. The next step is to make everything responsive. I added a feature to switch out the triangle video with a gif when the screen size gets smaller. All registration pages were added and the sponsor pages and links to their sites were added. I also tried to add a favico and it needs adjusted.