Final
Went in and deleted all the hide me and show me tags that we created for the triangle gifs to allow only one triangle per page. I adjusted some of the media queries so that my headings are centered and then get slightly smaller when the vw gets smaller. I fixed my register page so the buttons have space in between them and are not responsive. The sponsors page now has more space in between each logo and the the bottom of the footer so the image wasn't touching the footer. I still have some quirks like the heading not being completely centered in the triangle for all the sub-register pages, but everything else I am pretty happy with!